transcoding-service/requirements.txt
boto3
kafka-python
pydantic-settings
requests
